requirements.txt
aniso8601==7.0.0
arrow==1.2.2
asgiref==3.5.0
binaryornot==0.4.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
click==8.0.3
cookiecutter==1.7.3
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.12
django-allauth==0.47.0
django-browser-reload==1.3.0
django-cleanup==6.0.0
django-countries==7.2.1
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-dotenv==1.4.2
django-extensions==3.1.5
django-redis-sessions==0.6.2
django-tailwind==3.1.1
djangorestframework==3.13.1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.3
Jinja2==3.0.3
jinja2-time==0.2.0
MarkupSafe==2.0.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
poyo==0.5.0
promise==2.3
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.3
redis==4.1.3
requests==2.27.1
requests-oauthlib==1.3.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.8
Werkzeug==2.0.2
whitenoise==5.3.0
wrapt==1.13.3
